stub out most functions referring to pParent
pParent is a pointer to the “parent” VFS, that is the one registered
in sqlite as the previous VFS.

This is a way to delegate responsibility, but only works iff we have a
parent VFS in the first place (otherwise it will segfault).

Since we are not going to have a VFS like `unix` as parent, we have to
get rid of these. First we stub them out.

The important ones in fsOpen/fsDelete are not yet stubbed out.

The next step will be to pass a struct with all the function pointers
we need to call in these these stubs to sqlite3_onefile_from_c.