Give an example of passing the randomness to our VFS
I don’t know how to C!

But the genRandomness that we pass to sqlite seems to be used, so this
works!

We can now begin to implement all stuff still using pParent and
replace it with Ocaml runtime or solo5 primitives.

Also I added a printf to every stubbed out function, so it’s easy to
see what sqlite is trying to call internally.